move mail sending logic to the separate file
